Skip the hash property check/storage with an environment variable
This change adds a new environment variable
`DOCBOOK_TOOL_CONFLUENCE_SKIP_CONTENT_HASH_CHECKS`.

When enabled, it will skip the "content hash checks" that prevent making an API
request to Confluence to update the content when the content has not changed.
Note that this means every time the tool runs, the content will create a
version in Confluence, even if nothing has changed.

This is a workaround for some API issues with certain Confluence setups, so we
don't recommend enabling this unless you have a specific need to.